requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
reportlab==4.2.2
matplotlib>=3.8.0
pillow>=10.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
pandas>=2.0.0
